Use pg-type numeric for bigint in entities and use custom decimal in graph-node (#71)
* Use pg-type numeric for bigint columns in typeorm entities

* Use custom decimal implementation from util in graph-node

+++ packages/eden-watcher/src/entity/Distribution.ts
@@ -20,10 +20,10 @@ export class Distribution {
   @Column('varchar')
   distributor!: string;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   timestamp!: bigint;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   distributionNumber!: bigint;
 
   @Column('varchar')
