+++ packages/eden-watcher/src/entity/RewardScheduleEntry.ts
@@ -16,12 +16,12 @@ export class RewardScheduleEntry {
   @Column('integer')
   blockNumber!: number;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   startTime!: bigint;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   epochDuration!: bigint;
 
-  @Column('bigint', { transformer: bigintTransformer })
+  @Column('numeric', { transformer: bigintTransformer })
   rewardsPerEpoch!: bigint;
 }
